added PartsOnlyCheckpoint to prevent RuntimeErrors when pickling
After some experimentation determined that the model is checkpointing
fine and it is the pickling of the main loop that is causing the
program to halt with a RuntimeError. Since we don't necessarily
need the main loop saved, train-draw now uses a new subclass of
Checkpoint called PartsOnlyCheckpoint which works almost identicaly,
but does not attempt to pickle the main_loop.

Fixes jbornschein/draw#16