normalize datasets via DataStream.default_stream
fuel datasets have a set of default_transformations
that can be triggered by calling DataStream.default_stream.

This was tested on mnist, which caused the data to go
through its default transformer:

uint8_pixels_to_floatX(('features',))

and now results appear to match the results for bmnist.
This was also tested on bmnist where it had no effect
because that dataset has no default transformer.